refactor(api): move some test methods to parent class
